fix(layout): validate physical meter geometry

+++ src/main/java/fu/stockspace/stockspace_be/warehouse/dto/WarehouseBinResponse.java
@@ -18,12 +18,12 @@ public class WarehouseBinResponse {
     private BigDecimal maxWeight;
     private BigDecimal maxVolume;
     private Integer shelfLevel;
-    private Integer coordinateX;
-    private Integer coordinateY;
-    private Integer positionZ;
-    private Integer width;
-    private Integer length;
-    private Integer height;
+    private BigDecimal coordinateX;
+    private BigDecimal coordinateY;
+    private BigDecimal positionZ;
+    private BigDecimal width;
+    private BigDecimal length;
+    private BigDecimal height;
     private boolean isOccupied;
     private List<String> occupiedPositions;
 }
